refactor: split map viewmodel widget and watch support
