+++ scripts/ios_pr_check.sh
@@ -102,6 +102,7 @@ swift scripts/walk_return_to_origin_suggestion_unit_check.swift
 swift scripts/map_area_calculation_service_unit_check.swift
 swift scripts/map_log_unreachable_cleanup_unit_check.swift
 swift scripts/map_append_walk_point_discardable_unit_check.swift
+swift scripts/map_viewmodel_widget_watch_support_split_unit_check.swift
 swift scripts/map_auth_session_sync_unit_check.swift
 swift scripts/watch_context_update_gate_unit_check.swift
 swift scripts/watch_legacy_string_path_cleanup_unit_check.swift
